requirements.txt
python-dotenv
phidata
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
sqlalchemy
pgvector
psycopg[binary]
pypdf